# packages/order-service/openapi.yaml
openapi: 3.1.0
info:
  title: Order Service
  version: 1.0.0
paths:
  /orders:
    get:
      operationId: listOrders
      summary: List orders, optionally filtered by status
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [confirmed, rejected]
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      operationId: createOrder
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Insufficient stock or invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{id}:
    get:
      operationId: getOrder
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    CreateOrderRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
    Order:
      type: object
      required: [id, productId, quantity, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
        quantity:
          type: integer
        status:
          type: string
          enum: [confirmed, rejected]
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string